ofctrl.c: mff_ovn_geneve should be available at state WAIT_BEFORE_CLEAR.
When introducing the feature ovn-ofctrl-wait-before-clear, a new state
WAIT_BEFORE_CLEAR is added before CLEAR_FLOWS. In the WAIT_BEFORE_CLEAR
state the mff_ovn_geneve is already available. However, the function
ofctrl_get_mf_field_id() was not updated, and still returns 0 when the
state is WAIT_BEFORE_CLEAR, and will return the actual id when state
becomes CLEAR_FLOWS. This results in the I-P engine node
en_mff_ovn_geneve update, which triggers a recompute of the
physical_flow_output node, while this recompute could have happened
earlier in the WAIT_BEFORE_CLEAR state.

Although physical_flow_output recompute is not as heavy as
logical_flow_output recompute, it is still inefficient. This is a miss
from the ovn-ofctrl-wait-before-clear patch.

Fixes: 896adfd2d8b3 ("ofctrl: Support ovn-ofctrl-wait-before-clear to reduce down time during upgrade.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>