northd: Add lsp option force_fdb_lookup.
https://github.com/ovn-org/ovn/commit/f56ad41703d765de10aac3bbf74707e3ab2920c4

This option is applicable only if the lsp is of default 'type'
i.e. type=empty_string (which is a VM (or VIF) interface) and the
lsp also has 'unknown' addresses configured.
If lsp option 'force_fdb_lookup' is set to true, mac addresses
of the lsp (if configured) are not installed in the l2 lookup
table of the Logical Switch pipeline. However, MAC addresses
are learnt using the FDB Table.

Usecase:
=========
This option is required to reduce packet loss when VM is being
migrated across AZs (via VXLAN tunnel). lsp is configured in both
AZs on different logical switches which are sharing the same IP
subnet. If the port has unknown address set along with MAC, IP
then, any packet destined to VM's MAC on destination AZ will get
forwarded locally, which may lead to packet loss during VM migration.

This option is useful to reduce packet loss during VM migration
by forcing the logical switch to learn MAC using FDB.

The other way to address this issue is to use pkt_clone_type
option but this causes too much of packet duplication when there
are multiple ports with unknown address in the logical switch.

Signed-off-by: fanriming <fanriming@chinatelecom.cn>